Implement `Display`, `Error`, etc., for `Error`
Fixes #8.

In order to make the `cast` library more ergonomic, particularly when
used in conjuction with the `error-chain` crate (which provides
backtraces and easy error conversion), we define a number of new traits
for `cast::Error`:

- `Display`: This is required to implement `Error`.
- `Error`: This makes it possible to treat `cast::Error` as a regular
   Rust error.
- `Clone`: This makes it possible to duplicate a `cast::Error`.
- `Copy`: This gives "copy" semantics to `cast::Error`, allowing it to
  be automatically duplicated.  This is reasonable because `cast::Error`
  is a tiny enumeration where all the values are fixed.
- `Eq`: `cast::Error` already supported `PartialEq`, but since for all
  errors `e`, we know that `e == e`, it's also reasonable to implement
  `Eq`.

With the exception of `Error`, all of these can be implemented using
`core`.  For `Error`, we can only implement it if `std` is available,
which I have made optional using `feature="std"`.

This patch is a non-breaking API change.