travis: Test builds with and without `std`
Because we've added a new build option to this library, we should now
test it both ways.  Normally we would do this with `env:` and `matrix:`,
but that would make our `travis-cargo doc` builds trickier to enable
only on the right targets.  So we just test both ways in a single build
for simplicity.