Breaking change: Enable 128-bit integer support by default; increase MSRV.
Eliminate the need for a build script, and eliminate the rustc_version dependency,
to make cast.rs a more lightweight dependency for users. Increase the MSRV to 1.46.0
to match quickcheck 1.0.

Enable all the test cases for i128 and u128 according to the commented-out blocks. I
left two lines of test cases commented out since the tests fail to compile for them.

I documented the MSRV to be 1.46.0 because I found `cargo test` fails to build the
dependencies with 1.26.0, and it looks like quickcheck 1.0 has 1.46.0 as its MSRV.

I didn't upgrade the arbitrary dependency to 1.0 because some of the tests fail when
I tried. I filed an issue about this.