fix promote and back tests
These tests were *meant* to capture lossless two-way conversion between 2 types but older versions
of `quickcheck` did not provide enough coverage to verify that the intent was correct

When floats are involved in this test things get complicated because of lack of precision and
rounding

One failing test case was going from `4294967168_u32` to `f32` and back.
Using the built-in `as` operator on Rust 1.54.0 this results in:
- `4_294_967_168_u32` -> `4_294_967_300.0_f32` (NOTE this is larger than `u32::MAX` /
  `4_294_967_295`)
- `4_294_967_300.0_f32` -> `4_294_967_295_u32` (NOTE this is `u32::MAX`)

The behavior around this has changed recently; the last conversion used to be UB in older versions
of Rust.
With Rust 1.31.0 the last conversion returns `0_u32`

When using `cast::From` the `f32 -> u32` conversion returns `Err(Overflow)` so the test fails as
expected

Lossless conversion between integers and floats needs to consider the number of mantissa digits the
float type has:
- `f32` has 24 bits of mantissa so two-way lossless conversion is possible from `{i,u}8` and
  `{i,u}16`
- `f64` has 53 bits of mantissa so two-way lossless conversion is possible from `{i,u}8` and
  `{i,u}16` and `{i,u}32`

This commit updates the test to only test floats against integers than be losslessly converted into
said floats

NOTE we could retain the tests for e.g. `u32 -> f32` but we would need to clamp the input integer
value so that it doesn't exceed `1 << MANTISSA_DIGITS`, i.e. the two-way conversion is lossless