remove default Cargo features
Default features are pretty much impossible to disable in practice.

Cargo performs "union" of features across the dependency graph. If a single crate in the dep graph
enables the feature of a dependency then that feature cannot be disabled

The problem with default features is that is very likely they won't be checked at the moment of
adding a dependency. Meaning people will often enable features their library does *not* need.

Not having default features mean people will have to enable the features they need (= "pay for what
you use" in terms of compilation time, optional dependencies, etc.).